feat: Make the controller cursor snap to the "Done" button by default.
This change is made because initially snapping the cursor to the text box causes the left-click to fail to trigger the controller keyboard (the exact reason for this is unclear).
As a workaround, players will now need to move the cursor manually after opening the menu, after which they can click to open the virtual keyboard.